Add files via upload
Udacity Data Analyst Nanodegree Project: Analyze-AB-Test-Results

Description
In this project a dataset containing the conversion rates between new and old websites for an e-commerce website is going to be analyzed. This project is part of the Udacity Data Analysis Nanodegree course.

The data will be used to perform an A/B test, the goal is to help the company understand if they should implement the new page, keep the old page, or perhaps run the experiment longer to make their decision.